requirements.txt
proxmoxer
requests
python-dotenv
fastapi
uvicorn
redis
openai
google-genai
